enh: refactoring repo organization for adobe readers
