+++ nck/readers/__init__.py
@@ -32,14 +32,16 @@ from nck.readers.dbm_reader import dbm
 from nck.readers.dcm_reader import dcm
 from nck.readers.ga_reader import ga
 from nck.readers.search_console_reader import search_console
-from nck.readers.adobe_reader import adobe
-from nck.readers.adobe_reader_2_0 import adobe_2_0
 from nck.readers.radarly_reader import radarly
 from nck.readers.yandex_campaign_reader import yandex_campaigns
 from nck.readers.yandex_statistics_reader import yandex_statistics
 from nck.readers.gs_reader import google_sheets
 from nck.readers.confluence_reader import confluence
 from nck.readers.mytarget_reader import mytarget
+
+from nck.readers.adobe.cli import adobe
+from nck.readers.adobe_2_0.cli import adobe_2_0
+
 
 readers = [
     mysql,
