+++ nck/readers/adobe_2_0/helper.py
@@ -16,8 +16,9 @@
 # along with this program; if not, write to the Free Software Foundation,
 # Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
-from nck.config import logger
 from datetime import datetime
+
+from nck.config import logger
 
 
 class APIRateLimitError(Exception):
